refactor: remove import statement handling from facadeit_handler.py
